# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "HIMS APIs",
    "description": "APIs for Healthcare Information Management System(HIMS)."
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "multiparty/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/filter/roles/{role_id}/users": {
      "get": {
        "tags": [
          "Filters"
        ],
        "summary": "List all users with a given role-id",
        "description": "",
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/roles": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "List all roles",
        "description": "Get all roles",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Role"
        ],
        "summary": "Create a new role",
        "description": "Create a new role",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/roles/{id}": {
      "get": {
        "tags": [
          "Role"
        ],
        "summary": "Get a role by id",
        "description": "Get single role by role_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Role id."
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Roles retrieved successfull.",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Role"
        ],
        "summary": "Update a role",
        "description": "Edit single role",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "role_name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Role"
        ],
        "summary": "Delete a role by id",
        "description": "Delete single role by role_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "List all users",
        "description": "Get all users",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "description": "Create new user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User Info",
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by id",
        "description": "Get single user by user_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfull.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update a user",
        "description": "Edit single user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User Info",
            "schema": {
              "$ref": "#/definitions/EditUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a user by id",
        "description": "Delete single user by user_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "request-source-domain",
            "in": "header",
            "description": "Request Source Domain",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Role": {
      "type": "object",
      "properties": {
        "role_name": {
          "type": "string",
          "example": "Administrator"
        }
      },
      "required": [
        "role_name"
      ]
    },
...
    "AddUser": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        },
        "username": {
          "type": "string",
          "example": "jdoe"
        },
        "gender": {
          "type": "string",
          "example": "male"
        },
        "profile_image": {
          "type": "string",
          "example": "./public/images/usb-3.jpeg"
        },
        "dob": {
          "type": "string",
          "example": "2010-02-25"
        },
        "phone": {
          "type": "string",
          "example": "0789101112"
        },
        "role": {
          "type": "string",
          "example": "ewwe7sc6wefwc9sc7vf"
        },
        "email": {
          "type": "string",
          "example": "jdoe@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "password"
        },
        "status": {
          "type": "boolean",
          "example": true
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string",
              "example": "kingugi"
            },
            "city": {
              "type": "string",
              "example": "Dar es salaam"
            },
            "state": {
              "type": "string",
              "example": "Tanzania"
            },
            "postal_code": {
              "type": "string",
              "example": "123 Tza"
            }
          }
        }
      },
      "required": [
        "username",
        "role",
        "password"
      ]
    },
    "EditUser": {
      "type": "object",
      "properties": {
        "first_name": {
          "type": "string",
          "example": "John"
        },
        "last_name": {
          "type": "string",
          "example": "Doe"
        },
        "gender": {
          "type": "string",
          "example": "male"
        },
        "profile_image": {
          "type": "string",
          "example": "./public/images/usb-3.jpeg"
        },
        "dob": {
          "type": "string",
          "example": "2010-02-25"
        },
        "phone": {
          "type": "string",
          "example": "0789101112"
        },
        "role": {
          "type": "string",
          "example": "ewwe7sc6wefwc9sc7vf"
        },
        "email": {
          "type": "string",
          "example": "jdoe@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "password"
        },
        "status": {
          "type": "boolean",
          "example": true
        },
        "address": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string",
              "example": "kingugi"
            },
            "city": {
              "type": "string",
              "example": "Dar es salaam"
            },
            "state": {
              "type": "string",
              "example": "Tanzania"
            },
            "postal_code": {
              "type": "string",
              "example": "123 Tza"
            }
          }
        }
      }
    },
...
  }
}